Added AutoDetect Symlink To The Maker

+++ The Maker/maker.ps1
@@ -11,12 +11,16 @@ if (!(Test-Path $graphiteRoot)) {
     New-Item -Path $graphiteRoot -ItemType Directory | Out-Null
 }
 
-# Copy the folder
-Copy-Item -Path $TargetPath -Destination $destination -Recurse
+# If folder already exists in graphite
+if (Test-Path $destination) {
+    Write-Host "Folder already exists in graphite. Skipping copy."
+} else {
+    # Copy folder to graphite
+    Copy-Item -Path $TargetPath -Destination $destination -Recurse
+}
 
 # Remove the original folder
 Remove-Item -Path $TargetPath -Recurse -Force
 
-# Create a symlink
+# Create symbolic link
 cmd /c mklink /D `"$TargetPath`" `"$destination`"
-
